Model stacking for boosted trees

Model stacking with a two-layer architecture: first layer being boosted tree models fitted by xgboost, lightgbm, and catboost; second layer being a logistic regression model.

Usage

stackgbm(x, y, params, nfolds = 5L, seed = 42, verbose = TRUE)

Arguments

x

Predictor matrix

y

Response vector

params

A list of optimal parameters for boosted tree models. Can be derived from cv_xgboost, cv_lightgbm, and cv_catboost.

nfolds

Number of folds. Default is 5.

seed

Random seed for reproducibility

verbose

Show progress?

Value

Fitted boosted tree models and stacked tree model.

Examples

# check the vignette for code examples